+++ _scripts/teardown.sh
@@ -1,2 +1,2 @@
-docker-compose -f  docker-compose.dev.yaml down --volumes --remove-orphans
+docker-compose -f  docker-compose.dev.yaml -f  docker-compose.test.yaml down --volumes --remove-orphans
 docker volume prune --force