+++ client/static/js/lib/modal.js
@@ -11,6 +11,10 @@ all_modals.forEach((modal)=>{
 })
 
 
+//updatebuy
+function changeHoldingForm(id){
+    //
+
 // Close Modal 
 const modalClose = (modal) => {
     const modalToClose = document.querySelector('.'+modal);
@@ -22,7 +26,6 @@ const modalClose = (modal) => {
     let sect = document.getElementById('moreOptions');
     sect.classList.add('hidden');
 }
-
 
 //Open Modal 
 const openModal = (modal) => {
@@ -36,12 +39,19 @@ function showmoreoptions() {
     let sect = document.getElementById('moreOptions');
     sect.classList.remove('hidden');
 }
-
-
-
 // const { deleteSell, deleteBuy, updateBuy, createSell} = require('./requests');
 
 
+
+// onclick(deleteSellEvent(invest.id))
+function deleteSellEvent(data){
+    const p = document.getElementById('deleteMessage')
+    try{
+        deleteSell(data);
+        p.innerText = 'This past holding has been deleted';
+    }
+    catch(error){}
+}
 // //updatebuy
 // function changeHoldingForm(id){
 // asd
@@ -69,14 +79,13 @@ function showmoreoptions() {
 //     }
 //     catch(error)}
 
-    
-
-
-
-
-
+module.exports = {
+    changeHoldingForm,
+   // deletePastHoldingForm
+}
 
 // module.exports={
 //     changeHoldingForm,
 //     deletePastHoldingForm,
 // }
+